feat(utils): shared findProjectRoot under src/utils/fs
